Rotate cube along plane between all axes
Implement the Rotor type from geometric algebra to allow simple
rotations by the angle from one vector to another on the plane formed by
the vectors.

Use this to implement a cube rotation along the plane made by points
equidistant from all base axes.

Rotors will be fundamental to the rest of the game engine, such as
implementing a camera, so we introduce them here with an example usage.