Implement freelook camera
By adding camera rotations using the mouse, and rotating camera
movements to align with the direction the camera faces, we now have a
freelook camera which can be used to move around the world.

In addition, the C key allows looking backwards.

To allow seamless wrapping, confine the user's cursor the window and
hide it from view.