angle: Rename clamp function to wrap
Clamp is a misnomer for this operation, as we are wrapping the value
around the maximum, which has no effect on any calculations, beyond
keeping a higher precision.

A clamping function prevents the value from going beyond a certain
maximum, but does not wrap it around to the beginning.

Hence make the code clearer.