camera: Apply pitch followed by yaw
Until now, I believed that the camera should first yaw around its axis
and then pitch. However, if the camera pitches up and then spins 180
degrees, this would cause it to look down and in the opposite direction.

Instead, the camera should look in the opposite direction but remain
facing upwards, which can be achieved by applying the yaw after the
pitch.

With this change, our first-person freelook camera and movements finally
behave as expected.