camera: Use angle-plane rotations
Simplify yaw and pitch calculations by switching from angle-axis to
angle-plane calculations.

Introduce a BiVec3 type to represent a 3D bivector, or in other words,
the plane in the angle-plane calculations. Move the wedge implementation
from Vec3 to BiVec3, to avoid overloading the Vec3 type's applications,
and give more meaningful names to the components in the calculation.

Remove the axis-angle calculation code completely, including tests.