Add a skybox
Clear the screen to black and render a skybox around the player for Y=0
and above.

Use the Minecraft plains biome sky colour for the skybox.

Since the skybox renders for positive Y locations, and negative Y
locations remain black, there is now a visual indication of where up and
down are as you rotate the camera in the world.