textures: Switch from horizontal to vertical mesh
Until now we used a horizontal mesh for the cube texture, where the
bottom face's texture appeared to the right of the right face. Instead,
use a vertical mesh where the bottom face appears below the front face,
such that the rear / top / front / bottom faces appear in a vertical
line in the 2nd column of the mesh texture.

Vertical meshes seem slightly easier to reason about orientation when
drawing, so make this change early on before we introduce additional
textures.