Add a second render pass for UI elements
To implement the first UI element, the crosshair, we need to sample from
the framebuffer to ensure that we render a colour that contrasts well
with the scene.

Unlike Vulkan, WebGPU does not have subpasses that can sample from the
framebuffer. Instead, we render to an offscreen texture in the main
render pass, which we can later sample from.

To ensure that the contents of the offscreen render pass are visible in
the framebuffer, perform a "fullscreen copy" draw operation which
renders a full-screen quad, textured with the framebuffer itself.