Render a crosshair
Use the UI render pass to render a crosshair on the screen.

Most notably, the crosshair is represented as a grayscale texture,
currently of size 11x11 pixels, however the size can be varied. The
texture represents the crosshair's transparency. 0 (black) values
represent regions where the crosshair is not rendered, and anything
above represents regions where the crosshair is rendered with the
specified transparency value. A transparency of 0.75 was selected for
the base crosshair.

Since the crosshair is stored as a texture, users can customize their
crosshair however they please - rendering emojis, text, icons, or
anything else.

Note that, the crosshair size on-screen snaps to a multiple of the
texture size, so large texture sizes would be less flexible on various
resolutions.

The crosshair shader also samples from the current frame (produced by
the main render pass) at the locations where it renders, so that if the
colour of a specific pixel has components that are close to
1.0 (similar to the crosshair), the crosshair render inverts the
background colour to increase contract and thereby retain visibility.