wasm: Switch to cross-platform image reads [4 / 8]
Provide a platform-specific implementation to read a game resource as
bytes (`Vec<u8>`), and then layer an image decoding function on top of
that.

As a result, enabling image reading for wasm targets will be as simple
as providing a separate implementation which makes a network request for
the resource instead of the default desktop implementation which reads
from disk.