wasm: Extract desktop entrypoint [6 / 8]
Convert the project into a library, and extract a desktop binary which
uses a platform-specific entrypoint file to invoke the library
entrypoint (`start()`).

As a result, the code is no longer coupled to a desktop entrypoint and
can now be compiled for WASM.

Note that the library name is prefixed with an 'l' to avoid conflicting
build artefacts.