wasm: Reduce limits to WebGL 2 defaults [1 / 8]
To ensure that the game can run on WebGL 2, reduce device limits to the
WebGL 2 defaults, which are also the lowest defaults available. So far
we are not exceeding these, so we can keep this change
platform-agnostic.

To avoid restricting the size of the user's window, however, push the
texture resolution limits as high as the adapter allows.