wasm: Switch to cross-platform event loop [3 / 8]
On the web platform, the window (canvas) is not initialized until some
time after the event loop starts running. As a result, the window size
is reported as (0, 0) and we cannot create a graphics context since our
render surface size will also be invalid ((0, 0)).

To avoid generating errors in these cases, continously delay creation of
the graphics context until the window size becomes valid, and then
finally initialize the application.

Note that some other platforms may also require this approach, so the
implementation is not web-specific.

To facilitate this implementation, provide platform-specific
implementations for the following:
 - `init_window_attributes`: to allow the web context to map the
   relevant canvas to the window (no-op for desktop)
 - `run_future`: to allow the web context to schedule futures on its
   runtime (note that making these blocking is not possible on the web,
   so we switch to a non-blocking model)
 - `sleep`: to allow the web context to wait while the window is
   initialized