wasm: Avoid stripping release binaries [8 / 8]
Stripping information from the binary causes issues when running
`wasm-opt`, as the optimizer has no knowledge of the features supported
by the toolchain in that case, and ends up producing bulk memory
operations which are rejected by `wasm-validator`.

Originally we used this flag to reduce binary size. On Linux this causes
a binary size increase of 25%, however this only adds up to 2MB, which
is fine for the time being.