Automate discovery of assets folder in debug
In debug, instead of copying the assets folder to the build directory as
part of a build script, have the game search for the assets folder in
all parent directories of the running executable, eventually finding the
original copy part of the source tree.

This approach should save some cycles during builds (running a build
script and file copies) whilst also easing the development workflow (no
build script issues or dependency updates needed).

In release, retain the approach of only picking up the assets folder
located in the same directory as the executable to avoid parsing the
user's system (or accidentally picking up a different folder).

To achieve the above, we introduced a struct for resource reads which
eagerly picks out the assets folder on creation and errors if one does
not exist. This also helps out with the web implementation, allowing the
window origin to only be queried once and then remembered for all future
reads.