Improve cursor lock functionality
Ensure that the cursor is correctly locked across all platforms.

Using winit, attempt to confine the cursor (as before), but failing
that, also attempt to lock it, which provides this behaviour on some
additional platforms. Failing that yet again, implement a manual cursor
lock, always resetting the cursor to the center of the window when it
moves.

On web specifically, avoid doing any of the above as it does not work -
but when the user enters fullscreen, also lock their cursor as a
side-effect, ensuring we provide the proper full screen experience. Note
that the cursor remains locked even when they exit full screen, however
adding a special key just to lock the cursor seems cumbersome to explain
and remember.

Locking the cursor when entering fullscreen works because the browser
has detected interaction from the user to trigger the behaviour, but it
only works if we apply the full screen logic and cursor locking logic
during separate events (key down and key up), otherwise the entire
routine is considered too "long-running", which is indicative of
malicous webpages, and is rejected.

Note that the manual lock is possibly only necessary on X11, so
performance could be improved by conditionally compiling this away on
platforms such as Windows. However, this may change over time, with
winit adding support for X11 or adding additional window managers where
manual locking is necessary, so we choose to keep the code identical
across all non-web platforms for simplicity.