Implement chunking and flat terrain generation
Implement a terrain and chunking system, and augment the renderer to
allow the game to render chunks in a circular radius around the player,
which are loaded and unloaded as the player moves to simulate infinite
terrain in all cardinal directions.

Note that the model must now store chunk and block changes since the
last frame so that the renderer can update cached resources accordingly.
To facilitate this, we add an additional step to the composition loop
where we clear the changes from the model, as there is no other existing
place where this operation can be performed.

We also add information about the current block that the player is
looking at to the debug overlay.