math: Add Direction type
Use Direction to represent generic directions, as we currently only have
BoxFace to represent the direction in which a box face is facing.

In particular, we will initially be using this type to represent in
which direction the player has placed a block relative to the block they
are looking at.

Due to the inherent similarity between BoxFace and Direction, we add a
BoxFace -> Direction conversion and offload the normal operation to
Direction.