Add lossless and lossy conversion traits for types
Organise type conversions into three types:
 - infallible (`.into()`)
 - lossless
 - lossy

Introduce two new traits, `Coerce` and `CoerceLossy` to support the last
two, respectively.

Lossless conversions are those where we change numerical representation
but expect the number to remain exactly the same. We use
`.try_into().unwrap()` for this purpose, meaning we panic if these
conversions cannot be represented in the target type - which is fine, as
this should never happen and we should know if it does. We will mainly
use these conversions when converting from larger to smaller integers,
and between signed and unsigned integers.

Lossy conversions are those where we expect and agree that the new
numerical representation may lose some data or precision, particularly
when converting to float.

Conversions should be performed with `.into()` where possible, followed
by `.coerce()` (intended lossless conversion - checked at runtime), and
by `.coerce_lossy()` as a last resort.

By re-organising conversions under this scheme, we ensure that as many
conversions as possible are performed losslessly by invoking
`.try_into()`, whereas previously we may have used `as`.