Add font loading and text rendering
Use rusttype to load glyphs from a TTF file and render them on-screen.

Use the new text rendering functionality to display the current mean FPS
and frame time in a new debug menu.

The menu is enabled by default in debug builds and disabled in release
builds, but can always be toggled with F3.

We have elected to use the Monocraft font for text rendering in
Opencraft, which has been included in the assets submodule at version
4.2.1.

The Monocraft TTF file has been stripped of non-ASCII characters and
compressed to WOFF2 to reduce storage size in the git repo and reduce
startup latency for web users with a slow network connection. We use the
wuff library to decode the original TTF file from the WOFF2 file.

Note that only characters in active use are rasterised to reduce
computation and memory usage. As new characters are needed, they can
easily be added to the font atlas. Characters not already in the font
atlas will be converted to a '\0', which will be rasterised as a
rectangle, indicating a missing character.