Store colour attachment of final render pass
We started discarding the colour attachment in commit [1], however
whilst this works for the depth attachment, we need to store the result
of the colour attachment to display it on the screen.

Even though this was an incorrect usage of the APIs, most platforms
still displayed the result correctly, and the issue was only noticed
when running with the OpenGL ES and Vulkan drivers on Nvidia GPUs.

[1] e59b641c09ff2411dc9e7d4da8184e97f94942ae