Add dirt, stone, and bedrock blocks
Add the fundamental Minecraft blocks.

Add support for multiple block textures to the texture atlas.

Now that we have the necessary blocks, adjust the ClassicFlat generator
to produce the same blocks as in Minecraft - a layer of grass, two
layers of dirt, and a final layer of bedrock.

To showcase all of the different available blocks, now place stone
instead of grass when the player places a block.