blockgen: Add block generator program
Create a program to generate block textures from specifications encoded
in TOML.

To start with, create an algorithm to generate what are termed "plain"
blocks, such as dirt and sand. These blocks are generated by specifying
a palette of colours, with the absolute frequency that each colour
should occur in the texture. A random shuffle is then applied to
disperse the colours randomly within the texture pixels, creating a
blocky feel.

A deterministic random number generation algorithm, ChaCha8, was
intentionally selected to ensure textures can be reproduced by feeding
the same random seed to the generator (as long as the random operations
performed and their ordering remains unchanged within the program source
code), so the seeds are also (optionally) encoded in the TOML
specifications.